+++ bn-apps/cordapp-updates-distribution/corda-updates-app/src/test/kotlin/net/corda/businessnetworks/cordaupdates/app/ReloadConfigurationFlows.kt
@@ -1,0 +1,29 @@
+package net.corda.businessnetworks.cordaupdates.app
+
+import co.paralleluniverse.fibers.Suspendable
+import net.corda.businessnetworks.cordaupdates.app.member.MemberConfiguration
+import net.corda.businessnetworks.cordaupdates.transport.flows.RepositoryHosterConfigurationService
+import net.corda.core.flows.FlowLogic
+import net.corda.core.flows.StartableByRPC
+import java.io.File
+
+@StartableByRPC
+class ReloadMemberConfigurationFlow(private val fileName : String) : FlowLogic<Unit>() {
+    @Suspendable
+    override fun call() {
+        val config = File(ReloadMemberConfigurationFlow::class.java.classLoader.getResource(fileName).toURI())
+        val configuration = serviceHub.cordaService(MemberConfiguration::class.java)
+        configuration.reloadConfigurationFromFile(config)
+    }
+}
+
+@StartableByRPC
+class ReloadBNOConfigurationFlow(private val fileName : String) : FlowLogic<Unit>() {
+    @Suspendable
+    override fun call() {
+        val config = File(ReloadBNOConfigurationFlow::class.java.classLoader.getResource(fileName).toURI())
+        val configuration = serviceHub.cordaService(RepositoryHosterConfigurationService::class.java)
+        configuration.reloadConfigurationFromFile(config)
+    }
+}
+
