+++ bn-apps/cordapp-updates-distribution/corda-updates-core/build.gradle
@@ -1,13 +1,3 @@
-buildscript {
-    ext.slf4j_version = '1.7.25'
-    ext.tinylog_version = '1.3.5'
-    ext.typesafe_version = '1.3.2'
-
-    dependencies {
-        classpath "org.jetbrains.kotlin:kotlin-noarg:$kotlin_version"
-    }
-}
-
 apply plugin: 'kotlin'
 apply plugin: 'net.corda.plugins.cordapp'
 apply plugin: 'net.corda.plugins.quasar-utils'
@@ -18,7 +8,6 @@ cordapp {
         vendor "com.r3"
     }
 }
-
 
 sourceSets {
     main {
@@ -48,22 +37,13 @@ dependencies {
     }
 
     // transport over flows
-    compile project(':bn-apps:cordapp-updates-distribution:corda-updates-transport')
-
-    // typesafe
-    compile "com.typesafe:config:$typesafe_version"
+    cordapp project(':bn-apps:cordapp-updates-distribution:corda-updates-transport')
 
     cordaCompile "$corda_release_group:corda-core:$corda_release_version"
 
     testCompile "org.jetbrains.kotlin:kotlin-test:$kotlin_version"
     testCompile "junit:junit:$junit_version"
     testCompile project(":bn-apps:cordapp-updates-distribution:corda-updates-test-utils")
-
-    // SLF4J and TinyLog - convenient, small and fast logging libraries.
-    compile "org.slf4j:slf4j-api:$slf4j_version"
-    compile "org.slf4j:jul-to-slf4j:$slf4j_version"
-    compile "org.tinylog:tinylog:$tinylog_version"
-    compile "org.tinylog:slf4j-binding:$tinylog_version"
 }
 
 tasks.withType(org.jetbrains.kotlin.gradle.tasks.KotlinCompile).all {
