+++ bn-apps/cordapp-updates-distribution/corda-updates-core/src/main/kotlin/net/corda/businessnetworks/cordaupdates/core/CordappSyncer.kt
@@ -2,10 +2,7 @@ package net.corda.businessnetworks.cordaupdates.core
 
 import com.typesafe.config.Config
 import com.typesafe.config.ConfigFactory
-import net.corda.businessnetworks.cordaupdates.core.SyncerConfiguration.Companion.getIntOrNull
-import net.corda.businessnetworks.cordaupdates.core.SyncerConfiguration.Companion.getStringOrNull
 import net.corda.core.serialization.CordaSerializable
-import net.corda.core.utilities.loggerFor
 import org.eclipse.aether.RepositoryListener
 import org.eclipse.aether.artifact.DefaultArtifact
 import org.eclipse.aether.transfer.TransferListener
@@ -87,15 +84,17 @@ data class SyncerConfiguration(
         val httpProxyPassword : String? = null,
         val cordappSources : List<CordappSource>) {
     companion object {
-        fun readFromFile(file : File) : SyncerConfiguration  = readFromConfig(ConfigFactory.parseFile(file)!!)
+        fun readFromFile(file : File) : SyncerConfiguration = readFromConfig(ConfigFactory.parseFile(file)!!)
 
         fun readFromConfig(conf : Config) : SyncerConfiguration {
-            val cordapps = conf.getObjectList("cordappSources").asSequence().map { it.toConfig()!! }.map { CordappSource(
-                    it.getString("remoteRepoUrl"),
-                    it.getStringList("cordapps"),
-                    it.getStringOrNull("httpUsername"),
-                    it.getStringOrNull("httpPassword")
-            ) }.toList()
+            val cordapps = conf.getObjectList("cordappSources").asSequence().map { it.toConfig()!! }.map {
+                CordappSource(
+                        it.getString("remoteRepoUrl"),
+                        it.getStringList("cordapps"),
+                        it.getStringOrNull("httpUsername"),
+                        it.getStringOrNull("httpPassword")
+                )
+            }.toList()
 
             return SyncerConfiguration(
                     conf.getString("localRepoPath"),
