+++ bn-apps/cordapp-updates-distribution/corda-updates-transport/build.gradle
@@ -38,11 +38,9 @@ dependencies {
     // corda dependencies
     cordaCompile "$corda_release_group:corda-core:$corda_release_version"
     cordaCompile "$corda_release_group:corda-node-api:$corda_release_version"
-    cordaCompile "$corda_release_group:corda:$corda_release_version"
-    compile "$corda_release_group:corda-rpc:$corda_release_version"
+
     testCompile "$corda_release_group:corda-node-driver:$corda_release_version"
     testCompile project(":bn-apps:cordapp-updates-distribution:corda-updates-test-utils")
-
 
     // test dependencies
     testCompile "org.jetbrains.kotlin:kotlin-test:$kotlin_version"
