+++ bn-apps/cordapp-updates-distribution/corda-updates-transport/src/test/resources/corda-updates-with-allow-filter.conf
@@ -1,4 +1,4 @@
 repositories {
   default = "file:../TestRepo"
 }
-sessionFilter = net.corda.cordaupdates.transport.AllowAllSessionFilter
+sessionFilter = net.corda.businessnetworks.cordaupdates.transport.AllowAllSessionFilter
