+++ bn-apps/cordapp-updates-distribution/corda-updates-transport/src/test/resources/corda-updates-with-deny-filter.conf
@@ -1,4 +1,4 @@
 repositories {
   default = "file:../TestRepo"
 }
-sessionFilter = net.corda.cordaupdates.transport.DenyAllSessionFilter
+sessionFilter = net.corda.businessnetworks.cordaupdates.transport.DenyAllSessionFilter
